Jira: Use OAuth utils
