+++ extensions/jira/package.json
@@ -109,7 +109,7 @@
   ],
   "dependencies": {
     "@raycast/api": "^1.64.3",
-    "@raycast/utils": "^1.10.1",
+    "@raycast/utils": "^1.12.4",
     "date-fns": "^2.30.0",
     "filesize": "^10.1.0",
     "form-data": "^4.0.0",
