+++ extensions/jira/src/components/FormUserDropdown.tsx
@@ -2,10 +2,10 @@ import { Form, Icon } from "@raycast/api";
 import { useCachedPromise } from "@raycast/utils";
 import { ForwardedRef, forwardRef, useState } from "react";
 
+import { getJiraCredentials } from "../api/jiraCredentials";
 import { autocomplete } from "../api/request";
 import { User } from "../api/users";
 import { getUserAvatar } from "../helpers/avatars";
-import { getJiraCredentials } from "../helpers/withJiraCredentials";
 
 type FormUserDropdownProps = {
   autocompleteUrl?: string;
